tweak `local_bgs_refactored` file organization
